pyproject.toml
[tool.poetry]
name = "twittaloader"
version = "0.1.0"
description = ""
authors = ["Kenneth V. Domingo <hello@kvdomingo.xyz>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
requests = "^2.27.1"
python-dotenv = "^0.20.0"
loguru = "^0.6.0"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
pyinstaller = "^5.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']